requirements.txt
cryptography==44.0.2
fastapi==0.116.1
openai>=1.0
prometheus-client==0.21.1
pydantic-settings==2.10.1
python-dotenv>=1.0
python-multipart==0.0.20
PyJWT>=2.8
uvicorn==0.35.0

